[latest] publish built-ins to Open VSX registry
Added /src/publish-vsix.js to permit publishing to an Open VSX registry.
First run yarn to build the builtin extensions as .vsix in folder 'dist'
and remove any not wished published (e.g. not working well).

Then run:
   $> yarn publish:latest

To publish to a specific registry, use environment variable
OVSX_REGISTRY_URL to point to it. Defaults to open-vsx.org

Use environment variable OVSX_PAT to store your access token for the
registry.

Also improved the auto-generated README used in our .vsix, so it's clearer
that Microsoft is not involved in any way other than being the author of
these extensions we package and publish.

Added a copy of the VS Code license file for each builtin.

For vsix packaging, now use version number that reflects the VS Code
version as per its package.json.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>